Torch.Sig.Types.NN (signature[?])

Description

Global types used in Torch's Neural Network libraries (THNN and THCUNN).

FIXME: most of these are aliases and can either be removed in favour of Types or, in the case of CDim, should be pushed upstream back into Torch itself.

Documentation

type CNNState = CState

data CNNGenerator

data CDim